Make breakpad-symbols HTTP symbol retrieval optional
