+++ breakpad-symbols/Cargo.toml
@@ -7,11 +7,17 @@ license = "MIT"
 readme = "README.md"
 homepage = "https://github.com/rust-minidump/rust-minidump"
 repository = "https://github.com/rust-minidump/rust-minidump"
-exclude = [ "testdata/*" ]
+exclude = ["testdata/*"]
 edition = "2018"
 
 [badges]
 travis-ci = { repository = "rust-minidump/rust-minidump" }
+
+[features]
+# Private API, only here to support the fuzzer
+fuzz = []
+# Allow retrieval of symbols via HTTP
+http = ["reqwest", "tempfile"]
 
 [dependencies]
 async-trait = "0.1.51"
@@ -21,13 +27,13 @@ log = "0.4.1"
 minidump-common = { version = "0.10.4", path = "../minidump-common" }
 nom = "~1.2.2"
 range-map = "0.1.5"
-reqwest = { version = "0.11.6", features = ["gzip"] }
-tempfile = "3.3.0"
+reqwest = { version = "0.11.6", default-features = false, features = [
+    "gzip",
+    "rustls-tls",
+], optional = true }
+tempfile = { version = "3.3.0", optional = true }
 thiserror = "1.0.30"
 
-# Private API, only here to support the fuzzer
-[features]
-fuzz = []
-
 [dev-dependencies]
-tokio =  { version = "1.12.0", features = ["full"] }
+tempfile = "3.3.0"
+tokio = { version = "1.12.0", features = ["full"] }
